Updated benchmark code
Removed "printf" from most of the benchmarks. It tends to
be rather large and this hurts the size comparions. I will add
a "printf" benchmark at a later date.

Made sure that each benchmark checks the result and
reports "[OR]" or "[FAIL]".

VSCode insisted to format the files.

Reduced the number of iterations for the "fact" benchmark.