Increase epsilon with pow benchmark
As Oschonrock noted, it is not fair to call a benchmark "failing"
just because the compiler isn't using the IEEE-754 standard.
After all, it's just an optional addition to the C standard.

Increased the epsilon for value comparison to compensate.

For the average user, it really doesn't matter. There aren't
many usages for doing scientific calculations on a C64.
Btw. which processor do the Voyagers have?