Add some note to the fact result of LLVM
On the one hand, it's impressive that LLVM solved this benchmark.
On the other hand, it is not representable. It is not a demonstration
of the performance of the generated code - there is no code left.

After considering all the pros and cons and watching talks about
performance testing, etc., I decided to leave it that way.

After all, it demonstrates how well the build chain is working. If the
optimizer can reduce a function to zero execution time, it should do
so. Of course, LLVM was lucky to figure that out, and it's rare for
real-life programs (with user interaction). However, all the other
compilers are optimizing as well.

For example, Oscar64, and the developer notes that in the
documentation that it optimizes the Dhrystone benchmark to a point
where it puts the 6502 on par with a 4 MHz 8088. It does not
figure out that the fact() benchmark can be reduced to nothing,
but it optimizes the Dhrystore benchmark much better than other
compilers - it's just not that obvious. If I were to punish
LLVM for its "luck", I would have to do the same to all the other
compilers and turn off all optimizations.

In the end, I'll leave it as is, and maybe in the future I'll add
"unoptimized" benchmarks.